requirements.txt
crawl4ai==0.5.0.post4
python-dotenv>=1.0.0
selenium-driverless==1.9.4
requests>=2.28.0
aiohttp>=3.8.0
beautifulsoup4>=4.11.0
tqdm>=4.64.0
